+++ src/lwg.rs
@@ -33,100 +33,11 @@ use ya_client::model::market::{
 use ya_client::web::WebClient;
 use zip::CompressionMethod;
 
-#[derive(Clone)]
-struct DistStorage {
-    url: Arc<str>,
-}
+mod negociator;
+mod storage;
 
-struct DistSlot {
-    upload_url: String,
-    download_url: String,
-}
-
-impl DistSlot {
-    fn url(&self) -> &str {
-        self.upload_url.as_str()
-    }
-
-    async fn download(&self, out_path: &Path) -> anyhow::Result<()> {
-        let c = awc::Client::new();
-
-        let mut response = c
-            .get(&self.download_url)
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("download json: {}", e))?;
-
-        let payload = response.take_payload();
-        let mut fs = OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .open(out_path)?;
-        Ok(payload
-            .for_each(|b| {
-                let bytes = b.unwrap();
-                fs.write_all(bytes.as_ref()).unwrap();
-                future::ready(())
-            })
-            .await)
-    }
-
-    async fn download_json<T: DeserializeOwned>(&self) -> anyhow::Result<T> {
-        let c = awc::Client::new();
-        let b = c
-            .get(&self.download_url)
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("download json: {}", e))?
-            .body()
-            .await
-            .map_err(|e| anyhow::anyhow!("download json: {}", e))?;
-
-        Ok(serde_json::from_slice(b.as_ref())?)
-    }
-}
-
-impl DistStorage {
-    fn new(storage_url: Arc<str>) -> Self {
-        let url = storage_url;
-        Self { url }
-    }
-
-    async fn upload_bytes(&self, prefix: &str, bytes: Vec<u8>) -> anyhow::Result<String> {
-        let c = awc::Client::new();
-        let id = uuid::Uuid::new_v4();
-        let upload_url = format!("{}upload/{}-{}", self.url, prefix, id);
-
-        let response = c
-            .put(&upload_url)
-            .content_length(bytes.len() as u64)
-            .content_type("application/octet-stream")
-            .send_body(bytes)
-            .await
-            .map_err(|e| anyhow::anyhow!("upload bytes: {}", e))?;
-
-        Ok(format!("{}{}-{}", self.url, prefix, id))
-    }
-
-    async fn upload_file(&self, path: &Path) -> anyhow::Result<String> {
-        self.upload_bytes("blob", std::fs::read(path)?).await
-    }
-
-    async fn upload_json<T: Serialize>(&self, obj: &T) -> anyhow::Result<String> {
-        let bytes = serde_json::to_vec_pretty(obj)?;
-        self.upload_bytes("json", bytes).await
-    }
-
-    async fn download_slot(&self) -> anyhow::Result<DistSlot> {
-        let id = uuid::Uuid::new_v4();
-        let upload_url = format!("{}upload/out-{}", self.url, id);
-        let download_url = format!("{}out-{}", self.url, id);
-        Ok(DistSlot {
-            upload_url,
-            download_url,
-        })
-    }
-}
+use negociator::*;
+use storage::{DistSlot, DistStorage};
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "kebab-case")]
@@ -266,213 +177,6 @@ fn build_demand(node_name: &str, wasm_url: &str) -> Demand {
         demand_id: Default::default(),
         requestor_id: Default::default(),
     }
-}
-
-struct AgreementProducer {
-    subscription_id: String,
-    api: MarketRequestorApi,
-    my_demand: Demand,
-    pending: Vec<oneshot::Sender<String>>,
-}
-
-impl Actor for AgreementProducer {
-    type Context = Context<Self>;
-
-    fn started(&mut self, ctx: &mut Self::Context) {
-        let _ = ctx.run_interval(Duration::from_secs(10), |act, ctx| {
-            if !act.pending.is_empty() {
-                let requestor_api = act.api.clone();
-                let subscription_id = act.subscription_id.clone();
-                let me = ctx.address();
-
-                let _ = ctx.spawn(
-                    async move {
-                        let events = match requestor_api
-                            .collect(&subscription_id, Some(8.0), Some(5))
-                            .await
-                        {
-                            Ok(v) => v,
-                            Err(e) => {
-                                log::error!("fail to get market events: {}", e);
-                                return;
-                            }
-                        };
-                        for event in events {
-                            let _ = me.send(ProcessEvent(event)).await;
-                        }
-                    }
-                    .into_actor(act),
-                );
-            }
-        });
-    }
-
-    fn stopped(&mut self, ctx: &mut Self::Context) {
-        let subscription_id = self.subscription_id.clone();
-        let api = self.api.clone();
-        ctx.wait(
-            async move {
-                if let Err(e) = api.unsubscribe(&subscription_id).await {
-                    log::error!("unsubscribe error: {}", e);
-                }
-            }
-            .into_actor(self),
-        );
-    }
-}
-
-struct ProcessEvent(RequestorEvent);
-
-impl Message for ProcessEvent {
-    type Result = ();
-}
-
-struct NewAgreement;
-
-impl Message for NewAgreement {
-    type Result = Result<String, anyhow::Error>;
-}
-
-impl Handler<NewAgreement> for AgreementProducer {
-    type Result = ActorResponse<Self, String, anyhow::Error>;
-
-    fn handle(&mut self, msg: NewAgreement, ctx: &mut Self::Context) -> Self::Result {
-        let (tx, rx) = oneshot::channel();
-        self.pending.push(tx);
-
-        ActorResponse::r#async(
-            async move {
-                let agreement_id = rx.await?;
-                Ok(agreement_id)
-            }
-            .into_actor(self),
-        )
-    }
-}
-
-impl Handler<ProcessEvent> for AgreementProducer {
-    type Result = MessageResult<ProcessEvent>;
-
-    fn handle(&mut self, msg: ProcessEvent, ctx: &mut Self::Context) -> Self::Result {
-        match msg.0 {
-            RequestorEvent::ProposalEvent {
-                event_date: _,
-                proposal,
-            } => {
-                log::debug!(
-                    "processing ProposalEvent [{:?}] with state: {:?}",
-                    proposal.proposal_id,
-                    proposal.state
-                );
-                if proposal.state.unwrap_or(ProposalState::Initial) == ProposalState::Initial {
-                    if proposal.prev_proposal_id.is_some() {
-                        log::error!(
-                            "Proposal in Initial state but with prev id: {:#?}",
-                            proposal
-                        );
-                        return MessageResult(());
-                    }
-                    let bespoke_proposal = match proposal.counter_demand(self.my_demand.clone()) {
-                        Ok(v) => v,
-                        Err(e) => {
-                            log::error!(
-                                "problem with proposal: {:?} from {:?}: {}",
-                                proposal.proposal_id,
-                                proposal.issuer_id,
-                                e
-                            );
-                            return MessageResult(());
-                        }
-                    };
-                    let provider_id = proposal.issuer_id.clone().unwrap_or_default();
-                    let requestor_api = self.api.clone();
-                    let subscription_id = self.subscription_id.clone();
-                    let f = async move {
-                        let new_proposal_id = match requestor_api
-                            .counter_proposal(&bespoke_proposal, &subscription_id)
-                            .await
-                        {
-                            Ok(v) => v,
-                            Err(e) => {
-                                log::error!("counter_proposal fail: {}", e);
-                                return;
-                            }
-                        };
-                        log::debug!("new proposal id = {} for: {}", new_proposal_id, provider_id);
-                    };
-                    let _ = ctx.spawn(f.into_actor(self));
-                } else {
-                    // Try to create agreement
-                    if self.pending.is_empty() {
-                        return MessageResult(());
-                    }
-                    let new_agreement_id = proposal.proposal_id().unwrap().clone();
-                    let new_agreement = AgreementProposal::new(
-                        new_agreement_id.clone(),
-                        Utc::now() + chrono::Duration::hours(2),
-                    );
-                    let me = ctx.address();
-
-                    let requestor_api = self.api.clone();
-                    let me = ctx.address();
-                    let mut slot = match self.pending.pop() {
-                        Some(slot) => slot,
-                        None => return MessageResult(()),
-                    };
-                    let _ = ctx.spawn(
-                        async move {
-                            if let Err(e) = async {
-                                let _ack = requestor_api.create_agreement(&new_agreement).await?;
-                                log::info!("confirm agreement = {}", new_agreement_id);
-                                requestor_api.confirm_agreement(&new_agreement_id).await?;
-                                log::info!("wait for agreement = {}", new_agreement_id);
-                                requestor_api
-                                    .wait_for_approval(&new_agreement_id, Some(7.879))
-                                    .await?;
-                                Ok::<_, anyhow::Error>(())
-                            }
-                            .await
-                            {
-                                log::error!("fail to negotiate agreement: {}", new_agreement_id);
-                                Err(slot)
-                            } else {
-                                log::info!("agreement = {} CONFIRMED!", new_agreement_id);
-                                let _ = slot.send(new_agreement_id);
-                                Ok(())
-                            }
-                        }
-                        .into_actor(self)
-                        .then(|r, act, _ctx| {
-                            if let Err(slot) = r {
-                                act.pending.push(slot);
-                            }
-                            fut::ready(())
-                        }),
-                    );
-                }
-            }
-            _ => {
-                log::warn!("invalid response");
-            }
-        }
-        MessageResult(())
-    }
-}
-
-async fn agreement_producer(
-    market_api: &MarketRequestorApi,
-    demand: &Demand,
-) -> anyhow::Result<Addr<AgreementProducer>> {
-    let subscription_id = market_api.subscribe(demand).await?;
-    log::info!("sub_id={}", subscription_id);
-    let producer = AgreementProducer {
-        subscription_id,
-        api: market_api.clone(),
-        my_demand: demand.clone(),
-        pending: Default::default(),
-    };
-
-    Ok(producer.start())
 }
 
 struct PaymentManager {
@@ -770,10 +474,14 @@ async fn try_process_task(
         }
 
         log::info!("activity {} state: {:?}", activity_id, state);
-        let results = activity_api
+        let results = match activity_api
             .control()
             .get_exec_batch_results(&activity_id, &batch_id, Some(60))
-            .await?;
+            .await {
+            Ok(v) => v,
+            Err(ya_client::Error::TimeoutError { .. }) => Vec::default(),
+            Err(e) => Err(e)?
+        };
 
         log::info!("batch results {:?}", results);
 
@@ -855,23 +563,28 @@ pub fn run(
         let node_name = "test1";
         let my_demand = build_demand(node_name, &image);
         let market_api: ya_client::market::MarketRequestorApi = client.interface()?;
-        let a = agreement_producer(&market_api, &my_demand).await?;
+
         let storage = DistStorage::new(storage_server);
         let output_tasks = merge_path_ref.join("tasks.json");
         let payment_man = allocate_funds_for_task(&payment_api, tasks.len()).await?;
 
-        let agreements = futures::future::join_all(tasks.into_iter().map(|t| {
-            process_task(
-                storage.clone(),
-                client.clone(),
-                payment_man.clone(),
-                a.clone(),
-                task_output_path.clone(),
-                merge_path_ref.clone(),
-                t,
-            )
-        }))
-        .await;
+        let agreements = {
+            let a = agreement_producer(&market_api, &my_demand).await?;
+            let agreements = futures::future::join_all(tasks.into_iter().map(|t| {
+                process_task(
+                    storage.clone(),
+                    client.clone(),
+                    payment_man.clone(),
+                    a.clone(),
+                    task_output_path.clone(),
+                    merge_path_ref.clone(),
+                    t,
+                )
+            }))
+                .await;
+            let _ = a.send(Kill).await;
+            agreements
+        };
 
         let mut tasks = Vec::new();
         for agr in agreements {
